Rename code folder to src, as it may cause conflicts during debug (code is a standard module in Python)
