+++ src/timer.py
@@ -1,0 +1,31 @@
+from .settings import *
+
+class Timer:
+	def __init__(self, duration, repeat = False, autostart = False, func = None):
+		self.duration = duration
+		self.start_time = 0
+		self.active = False
+		self.repeat = repeat
+		self.func = func
+		
+		if autostart:
+			self.activate()
+
+	def __bool__(self):
+		return self.active
+
+	def activate(self):
+		self.active = True
+		self.start_time = pygame.time.get_ticks()
+
+	def deactivate(self):
+		self.active = False
+		self.start_time = 0
+		if self.repeat:
+			self.activate()
+
+	def update(self):
+		if self.active:
+			if pygame.time.get_ticks() - self.start_time >= self.duration:
+				if self.func and self.start_time != 0: self.func()
+				self.deactivate()
